requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
cfgv==3.3.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
identify==2.4.8
importlib-metadata==4.10.1
kombu==5.2.3
nodeenv==1.6.0
packaging==21.3
platformdirs==2.4.1
pre-commit==2.17.0
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
redis==4.1.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
vine==5.0.0
virtualenv==20.13.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
